fix task e2e specs
